A better approach.
this reverts the part from the last commit where i changed the schema.
now the UserSSOId is converted into a json string and then stored in
cassandra.

i also introduce 'cs' from the string-conversion package for
converting arbitrary pairs of string types into each other.  i've been
itching to do that for a long time, but if anybody dislikes this i can
revert it.